stub struct for redis snapshot - move shared redis structs to base module, move shared repository structs to base module
