+++ src/repository/mod.rs
@@ -1,4 +1,7 @@
 use serde::{Deserialize, Serialize};
+use thiserror::Error;
+
+use crate::decider::Event;
 
 #[cfg(feature = "esdb")]
 pub mod esdb;
@@ -15,4 +18,47 @@ pub enum RepositoryVersion<V> {
     Exact(V),
     NoStream,
     StreamExists,
+}
+
+#[derive(Debug, Error)]
+pub enum VersionedRepositoryError<RepoErr, V> {
+    #[error("Version conflict {0:?}")]
+    VersionConflict(VersionDiff<V>),
+    #[error("Repository Error {0}")]
+    RepoErr(RepoErr),
+}
+
+#[derive(Debug)]
+pub struct VersionDiff<V> {
+    expected: RepositoryVersion<V>,
+    actual: RepositoryVersion<V>,
+}
+
+impl<V: Clone> VersionDiff<V> {
+    pub fn new(expected: RepositoryVersion<V>, actual: RepositoryVersion<V>) -> Self {
+        Self { expected, actual }
+    }
+
+    pub fn expected(&self) -> RepositoryVersion<V> {
+        self.expected.to_owned()
+    }
+
+    pub fn actual(&self) -> RepositoryVersion<V> {
+        self.actual.to_owned()
+    }
+}
+
+pub trait WithFineGrainedStreamId {
+    fn to_fine_grained_id(&self) -> String;
+    fn fine_grained_eq(&self, comp: &str) -> bool {
+        comp == self.to_fine_grained_id()
+    }
+}
+
+pub trait StreamIdFromEvent<Evt: Event>: Sized {
+    fn from(e: Evt) -> Self {
+        Self::event_entity_id_into(e.get_id())
+    }
+
+    fn event_entity_id_into(id: <Evt as Event>::EntityId) -> Self;
 }
